feat: Delegator Count for Validators
We've added a new API which uses our GraphQL Endpoint to query a list of
active delegators for a validator.
Endpoint: `/staking/delegators/:validator_address`

Also refactored the GraphQL Queries to use string templates (which are
multiline by nature) instead of using string concatenation. This
improves readability and reduces surface for syntactical errors

Signed-off-by: jay-dee7 <jasdeepsingh.uppal@gmail.com>