cast all numeric constants to Number again.
I decided to do this after noticing some requests like this:

https://api.cheqd.net/cosmos/staking/v1beta1/validators/cheqdvaloper1akqyyul8q32nzxa2yzkenex0q2l4dx6udpjfdf/delegations?pagination.count_total=false&pagination.limit=50&pagination.offset=050

As you can see offset is set to "050", which should be the case of our Numeric constants being are considered as String.

I highly doubt this was happing because we're wrapping our Numeric constants in quotes in toml file.